Support variable when seeking by property index in match clause (#5553)
* Support variable when seeking by prop index in match clause

* Rename

* Move the index query context construction outside of OptRule

* Move the optimizer utils to common graph util folder

* Eval argument values in index scan executor

* Fix bug

* refactor and fix error

* Update tck tests

---------
