+++ src/graph/planner/CMakeLists.txt
@@ -27,6 +27,7 @@ nebula_add_library(
     match/MatchPathPlanner.cpp
     match/ShortestPathPlanner.cpp
     match/VariableVertexIdSeek.cpp
+    match/VariablePropIndexSeek.cpp
     ngql/PathPlanner.cpp
     ngql/GoPlanner.cpp
     ngql/SubgraphPlanner.cpp
